requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.5.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1749054771823/work
annotated-doc==0.0.4
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio==4.11.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==26.1.0
bitsandbytes==0.45.5
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
certifi @ file:///croot/certifi_1752653701669/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
conda @ file:///croot/conda_1749171252582/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1751061994888/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1752801564376/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
contourpy==1.3.0
cryptography @ file:///croot/cryptography_1748977429425/work
cycler==0.12.1
debugpy==1.8.20
distro @ file:///croot/distro_1714488253808/work
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
faiss-gpu-cu12==1.12.0
fastapi==0.128.8
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.60.2
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2025.10.0
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==0.17.3
httpx==0.24.0
huggingface_hub==0.36.2
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
latex2mathml==3.78.1
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
# Editable install with no version control (llava==1.2.2.post1)
-e /data/test03/HIRE-main
markdown-it-py @ file:///croot/markdown-it-py_1684279902645/work
markdown2==2.5.5
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl @ file:///opt/conda/conda-bld/mdurl_1659716024347/work
menuinst @ file:///croot/menuinst_1750861706497/work
mpmath==1.3.0
narwhals==2.18.0
networkx==3.2.1
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
orjson==3.11.5
packaging @ file:///croot/packaging_1734472117206/work
panda==0.3.1
pandas==2.3.3
peft==0.17.1
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
protobuf==6.33.6
psutil==7.2.2
pycocotools==2.0.11
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1750768261732/work
pydantic_core @ file:///croot/pydantic-core_1750754736071/work
pydub==0.25.1
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing==3.3.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests @ file:///croot/requests_1750426329888/work
rich @ file:///croot/rich_1732638981168/work
rpds-py==0.27.1
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
ruff==0.15.7
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.37.2
triton==2.1.0
typer==0.23.2
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.3
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.39.0
wavedrom==2.0.3.post3
websockets==11.0.3
zipp==3.23.0
zstandard @ file:///croot/zstandard_1731356346222/work
